requirements.txt
numpy
pandas
keras
tensorflow
scikit-learn
pillow
matplotlib
pyttsx3